# Improve performance sort TPCH q3 with Utf8Vew ( Sort-preserving mergi…