@zhuqi-lucas 2025-03-28
After thinking more, i think the better way is to improve it in arrow-rs, so we will benefit more about the utf8view regression cases, created the ticket for arrow-rs:

https://github.com/apache/arrow-rs/issues/7350