@zhuqi-lucas 2025-03-28
First of all, i create the reproducer benchmark PR for arrow-rs:

https://github.com/apache/arrow-rs/pull/7351

| Benchmark                     | Utf8 Time (µs)    | Utf8View Time (µs) | Slower (%)              |
|--------------------------------|------------------|--------------------|------------------------|
| eq custom                      | 97.426 - 97.940  | 210.93 - 214.13    | 116%  |
| neq custom                     | 100.27 - 102.04  | 211.39 - 212.46    | 111%  |
| gt custom                      | 84.338 - 84.634  | 196.90 - 197.52    | 133%  |
| gt custom scalar               | 99.384 - 99.817  | 181.69 - 182.25    | 83%   |
| gt custom scalar long string   | 100.32 - 100.65  | 212.39 - 212.92    | 112%  |
| like custom scalar             | 89.111 - 90.811  | 177.80 - 180.97    | 100%  |
| nlike custom scalar            | 87.976 - 89.682  | 98.116 - 101.02    | 12%   |
| ilike custom scalar            | 199.82 - 200.68  | 312.25 - 314.25    | 56%   |
| nilike custom scalar           | 198.95 - 199.67  | 232.68 - 234.71    | 17%   |